requirements.txt
fastapi>=0.100.0
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
requests
httpx>=0.24.0
starlette>=0.27.0
cryptography
setfit
sentence-transformers
datasets
python-multipart
google-genai
